Rename new show markdown and split into own file.
