+++ lib/features/neural/display.dart
@@ -36,7 +36,7 @@ import 'package:rattle/utils/image_exists.dart';
 import 'package:rattle/utils/word_wrap.dart';
 import 'package:rattle/widgets/image_page.dart';
 import 'package:rattle/widgets/page_viewer.dart';
-import 'package:rattle/widgets/show_markdown_file.dart';
+import 'package:rattle/utils/show_markdown_file.dart';
 import 'package:rattle/widgets/text_page.dart';
 
 /// The panel displays the instructions or the output.
