+++ lib/features/transform/panel.dart
@@ -30,7 +30,7 @@ library;
 import 'package:flutter/material.dart';
 
 import 'package:rattle/constants/markdown.dart';
-import 'package:rattle/widgets/show_markdown_file.dart';
+import 'package:rattle/utils/show_markdown_file.dart';
 
 /// The SVM feature supports Svm rule analysis.
