+++ lib/features/visual/display.dart
@@ -37,7 +37,7 @@ import 'package:rattle/providers/vars/types.dart';
 import 'package:rattle/utils/image_exists.dart';
 import 'package:rattle/widgets/page_viewer.dart';
 import 'package:rattle/widgets/image_page.dart';
-import 'package:rattle/widgets/show_markdown_file.dart';
+import 'package:rattle/utils/show_markdown_file.dart';
 
 /// The panel displays the instructions or the output.
