+++ lib/features/wordcloud/display.dart
@@ -38,7 +38,7 @@ import 'package:rattle/r/extract.dart';
 import 'package:rattle/providers/stdout.dart';
 import 'package:rattle/widgets/image_page.dart';
 import 'package:rattle/widgets/page_viewer.dart';
-import 'package:rattle/widgets/show_markdown_file.dart';
+import 'package:rattle/utils/show_markdown_file.dart';
 import 'package:rattle/widgets/text_page.dart';
 
 class WordCloudDisplay extends ConsumerStatefulWidget {
